Separate AccountFlags to its own file
