+++ UT4MasterServer.Models/Database/Account.cs
@@ -5,20 +5,6 @@ namespace UT4MasterServer.Models.Database;
 using System.Text.Json.Serialization;
 using UT4MasterServer.Common.Helpers;
 
-[Flags]
-public enum AccountFlags
-{
-	None = 0,
-
-	Admin = 1,
-	Moderator = 2,
-	Developer = 4,
-	ContentCreator = 8,
-	HubOwner = 16,
-
-	All = ~0
-}
-
 [BsonIgnoreExtraElements]
 public class Account
 {
